add component props
